fix admin page error

+++ site/src/commonMain/kotlin/com/saitawngpha/blogmultiplatform/models/Post.kt
@@ -13,8 +13,8 @@ import com.saitawngpha.blogmultiplatform.models.Category
     @SerialName("_id")
      val id: String = "",
      val author: String = "",
-//     val date: Double = 0.0,
-     val date: Long = 0L,
+     val date: Double = 0.0,
+//     val date: Long = 0L,
      val title: String ,
      val subtitle: String ,
      val thumbnail: String ,
